CRUD - San Pham, validation form
validation form san pham - ma giam gia - danh muc - ma giam gia - nguoi dung - nha cung cap - phieu nhap

+++ src/main/resources/static/assets/admin/layout/js-ctrl/nguoidung-ctrl.js
@@ -47,48 +47,61 @@ app.controller("nguoidung-ctrl", function ($scope, $http, $location) {
     $scope.create = function () {
         var item = angular.copy($scope.form);
         var url = `${host}/nguoidung`;
-        $http.post(url, item).then((resp) => {
+        if(!$scope.myform.$valid) {
+            return;
+        } 
+        else {
+            $http.post(url, item).then((resp) => {
 
-            $scope.items.push(item);
-            $scope.reset();
-            $scope.initialize();  //reload lại items
-            console.log("Success", resp)
-            alert("Thêm thành công!")
-        }).catch((err) => {
-            alert("Lỗi thêm người dùng!")
-            console.log("Error", err)
-        });
+                $scope.items.push(item);
+                $scope.reset();
+                $scope.initialize();  //reload lại items
+                
+                console.log("Success", resp)
+                alert("Thêm thành công!")
+            }).catch((err) => {
+                alert("Lỗi thêm người dùng!")
+                console.log("Error", err)
+            });
+        }
+        
     }
 
     $scope.update = function () {
         var item = angular.copy($scope.form);
         var url = `${host}/nguoidung/${$scope.form.nguoiDungId}`;
-        $http.put(url, item).then((resp) => {
-            var index = $scope.items.findIndex(item => item.nguoiDungId == $scope.form.nguoiDungId);
-            $scope.items[index] = resp.data;
-            $scope.reset();
-            $scope.initialize();  //reload lại items
-            $(".nav-tabs button:eq(1)").tab("show"); //hiển thị bảng
-            console.log("Success", resp)
-            alert("Cập nhật thành công!")
-        }).catch((err) => {
-            console.log("Error", err)
-            alert("Lỗi cập nhật!")
-        });
+        if(!$scope.myform.$valid) {
+            return;
+        } else {
+            $http.put(url, item).then((resp) => {
+                var index = $scope.items.findIndex(item => item.nguoiDungId == $scope.form.nguoiDungId);
+                $scope.items[index] = resp.data;
+                $scope.reset();
+                $scope.initialize();  //reload lại items
+                $(".nav-tabs button:eq(1)").tab("show"); //hiển thị bảng
+                console.log("Success", resp)
+                alert("Cập nhật thành công!")
+            }).catch((err) => {
+                console.log("Error", err)
+                alert("Lỗi cập nhật!")
+            });
+        }
     }
 
     $scope.delete = function (nguoiDungId) {
         var url = `${host}/nguoidung/${nguoiDungId}`;
-        $http.delete(url).then((resp) => {
-            var index = $scope.items.findIndex(item => item.nguoiDungId == nguoiDungId);
-            $scope.items.splice(index, 1);
-            $scope.reset();
-            console.log("Success", resp)
-            alert("Xóa thành công!")
-        }).catch((err) => {
-            console.log("Error", err)
-            alert("Lỗi xóa người dùng!")
-        });
+        if(confirm('Xác nhận xóa?')) {
+            $http.delete(url).then((resp) => {
+                var index = $scope.items.findIndex(item => item.nguoiDungId == nguoiDungId);
+                $scope.items.splice(index, 1);
+                $scope.reset();
+                console.log("Success", resp)
+                alert("Xóa thành công!")
+            }).catch((err) => {
+                console.log("Error", err)
+                alert("Lỗi xóa người dùng!")
+            });
+        }
     }
