+++ src/main/resources/static/assets/admin/layout/js-ctrl/phieunhap-ctrl.js
@@ -62,48 +62,58 @@ app.controller("phieunhap-ctrl", function ($scope, $http, $location) {
     $scope.create = function () {
         var item = angular.copy($scope.form);
         var url = `${host}/phieunhap`;
-        $http.post(url, item).then((resp) => {
+        if (!$scope.myform.$valid) {
+            return;
+        } else {
+            $http.post(url, item).then((resp) => {
 
-            $scope.items.push(item);
-            $scope.reset();
-            $scope.initialize(); //reload lại items
-            console.log("Success", resp)
-            alert("Thêm thành công!")
-        }).catch((err) => {
-            alert("Lỗi thêm phiếu nhập!")
-            console.log("Error", err)
-        });
+                $scope.items.push(item);
+                $scope.reset();
+                $scope.initialize(); //reload lại items
+                console.log("Success", resp)
+                alert("Thêm thành công!")
+            }).catch((err) => {
+                alert("Lỗi thêm phiếu nhập!")
+                console.log("Error", err)
+            });
+        }
     }
 
     $scope.update = function () {
         var item = angular.copy($scope.form);
         var url = `${host}/phieunhap/${$scope.form.phieuNhapId}`;
-        $http.put(url, item).then((resp) => {
-            var index = $scope.items.findIndex(item => item.phieuNhapId == $scope.form.phieuNhapId);
-            $scope.items[index] = resp.data;
-            $scope.reset();
-            $scope.initialize();  //reload lại items
-            $(".nav-tabs button:eq(1)").tab("show"); //hiển thị bảng
-            console.log("Success", resp)
-            alert("Cập nhật thành công!")
-        }).catch((err) => {
-            console.log("Error", err)
-            alert("Lỗi cập nhật!")
-        });
+        if (!$scope.myform.$valid) {
+            return;
+        } else {
+            $http.put(url, item).then((resp) => {
+                var index = $scope.items.findIndex(item => item.phieuNhapId == $scope.form.phieuNhapId);
+                $scope.items[index] = resp.data;
+                $scope.reset();
+                $scope.initialize();  //reload lại items
+                $(".nav-tabs button:eq(1)").tab("show"); //hiển thị bảng
+                console.log("Success", resp)
+                alert("Cập nhật thành công!")
+            }).catch((err) => {
+                console.log("Error", err)
+                alert("Lỗi cập nhật!")
+            });
+        }
     }
 
     $scope.delete = function (phieuNhapId) {
         var url = `${host}/phieunhap/${phieuNhapId}`;
-        $http.delete(url).then((resp) => {
-            var index = $scope.items.findIndex(item => item.phieuNhapId == phieuNhapId);
-            $scope.items.splice(index, 1);
-            $scope.reset();
-            console.log("Success", resp)
-            alert("Xóa thành công!")
-        }).catch((err) => {
-            console.log("Error", err)
-            alert("Lỗi xóa danh mục!")
-        });
+        if (confirm('Xác nhận xóa?')) {
+            $http.delete(url).then((resp) => {
+                var index = $scope.items.findIndex(item => item.phieuNhapId == phieuNhapId);
+                $scope.items.splice(index, 1);
+                $scope.reset();
+                console.log("Success", resp)
+                alert("Xóa thành công!")
+            }).catch((err) => {
+                console.log("Error", err)
+                alert("Lỗi xóa danh mục!")
+            });
+        }
     }
 
 
@@ -138,6 +148,6 @@ app.controller("phieunhap-ctrl", function ($scope, $http, $location) {
         }
     };
 
-    
+
 
 })
